SynthDiag.jl

SynthDiag.jl

Installation

Synthetic Diagnostics

Interferometer

Langmuir Probes

Synthetic Actuators

Gas Injection

Noise model

Installation

SynthDiag is registered with public repository FuseRegistry. For installation:

using Pkg
Pkg.Registry.add(RegistrySpec(url="https://github.com/ProjectTorreyPines/FuseRegistry.jl.git"))
Pkg.Registry.add("General")
Pkg.add("SynthDiag")

Synthetic Diagnostics

Interferometer

Synthetic interferometer can be added using IMAS compatible JSON file that describes the metadata and the line of sight of chords. On computation, the interferometer uses edge profiles and core profiles data in the IMAS IDS to numerically integrate electron density along the line of sight and returns data in IMAS IDS interferometer object.

SynthDiag.add_interferometer! — Function

add_interferometer!(
    config::String=default_ifo,
    @nospecialize(ids::IMASDD.dd)=IMASDD.dd();
    overwrite::Bool=false, verbose::Bool=false, rtol::Float64=1e-3, n_e_gsi::Int=5,
)::IMASDD.dd

Add interferometer to IMAS structure using a JSON file and compute the line integrated electron density if not present

add_interferometer!(
    config::Dict{Symbol, Any},
    @nospecialize(ids::IMASDD.dd)=IMASDD.dd();
    overwrite::Bool=false, verbose::Bool=false, rtol::Float64=1e-3, n_e_gsi::Int=5,
)::IMASDD.dd

Add interferometer to IMAS structure using a Dict and compute the line integrated electron density if not present

SynthDiag.compute_interferometer! — Function

compute_interferometer!(
    @nospecialize(ids::IMASDD.dd);
    rtol::Float64=1e-3,
    n_e_gsi::Int=5,
)

Computed the line integrated electron density from the interferometer data present in IDS structure for all the chords. The computation is based on the edge profile data and core profile data present in the IDS structure.

Langmuir Probes

Langmuir probes can be added using IMAS compatible JSON file that describes the metadata and positions of embedded or reciprocating probes. Computation is currently supported for embedded langmuir probes only which uses the edge profiles data to report the edge electron and average ion temperature and electron density. If plasma potential and probe biasing information is available, it will use a langmuir probe current model to also calculate the ion saturaton current and current density as reported by a typical probe in IMAS data format.

SynthDiag.add_langmuir_probes! — Function

add_langmuir_probes!(
    config::String=default_lp,
    @nospecialize(ids::IMASDD.dd)=IMASDD.dd();
    overwrite=false, verbose=false, kwargs...,
)::IMASDD.dd

Add langmuir probes positions and other parameters from JSON file to ids structure

add_langmuir_probes!(
    config::Dict{Symbol, Any},
    @nospecialize(ids::IMASDD.dd)=IMASDD.dd();
    overwrite=false, verbose=false, kwargs...,
)::IMASDD.dd

Add langmuir probes positions and other parameters from Dictionary to ids structure

SynthDiag.compute_langmuir_probes! — Function

compute_langmuir_probes!(
    ids::IMASDD.dd;
    v_plasma::Union{Float64, Vector{Float64}, Nothing}=nothing,
    v_floating::Union{Float64, Vector{Float64}, Nothing}=nothing,
    v_probe::Union{Float64, Vector{Float64}, Nothing}=nothing,
    v_plasma_noise::Union{Noise, Nothing}=nothing,
    Z_eff::Float64=1.0,
    m_i_amu::Float64=2.014, # Deuterium
    probe_type::Symbol=:cylindrical,
    ne_noise::Union{Noise, Nothing}=nothing,
    te_noise::Union{Noise, Nothing}=nothing,
    ti_noise::Union{Noise, Nothing}=nothing,
    n_e_gsi::Int=5,
)

Compute langmuir probe outputs for all the embedded langmuir probes in the ids structure using edge profiles data. If v_plasma or v_floating are provided along with v_probe, then langmuir_probe_current() is used to compute the ion saturation current and parallel current density and stored in the IDS. Noise models can be added for electron density and temperature and ion average temperature as power spectral densities using data type Noise.

Input arguments:

ids: IMASDD.dd object

v_plasma: Plasma potential (in V) or a vector of plasma potentials for each time step

v_floating: Floating potential (in V) or a vector of floating potentials for each time step

v_probe: Probe potential (in V) or a vector of probe potentials for each time step

v_plasma_noise: Noise object for plasma potential

Z_eff: Effective charge of the ion

m_i_amu: Ion mass in atomic mass units (amu)

probe_type: Type of the probe, either :cylindrical or :spherical

ne_noise: Noise object for electron density

te_noise: Noise object for electron temperature

ti_noise: Noise object for average ion temperature

n_e_gsi: Grid subset index that is used in edge profiles data for electron density, electron temperature, and average ion temperature.

SynthDiag.langmuir_probe_current — Function

langmuir_probe_current(
    Δv::Float64,
    Te::Float64,
    Ti=::Float64,
    ne::Float64,
    A::Float64,
    m_i_amu=2.04,
    Z_eff::Float64=1.0;
    probe_type::Symbol=:cylindrical,
)::Float64

Langmuir Probe current model for cylindrical and spherical probes.

For Δv <= 0:

$I_{probe} = -I_{isc} (1 - \frac{Δv}{T_i})^{ex} - I_{esc} e^{\frac{Δv}{T_e}}$

For Δv > 0:

$I_{probe} = -I_{esc} (1 + \frac{Δv}{T_e})^{ex} - I_{isc} e^{-\frac{Δv}{T_i}}$

where

$I_{isc} = \frac{1}{2} A q_i n_e \sqrt{e (T_e + T_i) / m_i}$

$I_{esc} = -\frac{1}{4} A e n_e \sqrt{8 e T_e / m_e}$

where A is the effective surface area of the probe, q_i(Z_effe) is the total charge of the ion, and ex is 1 for spherical and 0.5 for cylindrical probes.

Input arguments:

Δv: Bias voltage (v_probe - v_plasma)

Te: Electron temperature (in eV)

Ti: Ion temperature (in eV)

ne: Electron density (in $m^{-3}$)

A: Effective surface area of the probe (in $m^{2}$)

m_i_amu: Ion mass in atomic mass units (amu)

Z_eff: Effective charge of the ion (in units of elementary charge e)

probe_type: Type of the probe, either :cylindrical or :spherical

Ref: L. Conde, "An introduction to Langmuir probe diagnostics of plasmas" (2011)

Synthetic Actuators

Gas Injection

Gas valves can be added using IMAS compatible JSON file that describes the metadata, response curve, and positions of gas valves. If ids.gas_injection.valve[:].voltage.data is present, the gas output flow rate is calculated. Additionally, if a valve model dictionary is passed, more realistic actuation can be modeled that includes second order low pass effect, latency, and dribble effect.

SynthDiag.add_gas_injection! — Function

add_gas_injection!(
    config::String=default_gas_injection,
    @nospecialize(ids::IMASDD.dd)=IMASDD.dd();
    overwrite::Bool=false,
    verbose::Bool=false,
)::IMASDD.dd

Add gas valves from a JSON file and compute the gas flow rate based on the command signal in the gas valves.

add_gas_injection!(
    config::Dict{Symbol, Any},
    @nospecialize(ids::IMASDD.dd)=IMASDD.dd();
    overwrite::Bool=false,
    verbose::Bool=false,
)::IMASDD.dd

Add gas valves from a dictionary and compute the gas flow rate based on the command signal in the gas valves.

SynthDiag.compute_gas_injection — Function

compute_gas_injection(
    tt::Vector{Float64},
    cmd_voltage::Vector{Float64},
    response_curve_voltage::Vector{Float64},
    response_curve_flow_rate::Vector{Float64};
    valve_model::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    global_latency::Float64=0.0,
)::Tuple{Vector{Float64}, Vector{Float64}, Vector{Float64}}

Lowest level function to compute gas flow rate based on the command voltage data and response cruve data all provided in base Julia types.

compute_gas_injection(
    tt::Vector{Float64},
    cmd_voltage::Vector{Float64},
    response_curve::IMASDD.gas_injection__valve___response_curve;
    valve_model::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    global_latency::Float64=0.0,
)::Tuple{Vector{Float64}, Vector{Float64}, Vector{Float64}}

Convinience function format where response curve is provided as the ids type but everything else is provided in base Julia types.

compute_gas_injection(
    valve::IMASDD.gas_injection__valve;
    valve_model::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    global_latency::Float64=0.0,
)::Tuple{Vector{Float64}, Vector{Float64}, Vector{Float64}}

Top most level function to compute gas flow rate for a single valve.

SynthDiag.compute_gas_injection! — Function

compute_gas_injection!(
    ids::IMASDD.dd;
    valves::Dict{String, Dict{Symbol, Any}}=Dict{String, Dict{Symbol, Any}}(),
)::Array{Vector{Float64}}

Compute the gas flow rate based on the command signal in the all gas valves.

compute_gas_injection!(
    valve::IMASDD.gas_injection__valve;
    valve_model::Dict{Symbol, Any}=Dict{Symbol, Any}(),
    global_latency::Float64=0.0,
)::Vector{Float64}

In-place version of compute_gas_injection function for a single valve.

SynthDiag.get_lpf — Function

get_lpf(
    fs::Float64,
    tau::Float64,
    damping::Float64,
    gain::Float64,
)::SecondOrderSections

Create a second order filter with the given parameters. The filter is created in the s-domain and then converted to the z-domain using bilinear transform. In s-domain, the filter transfer function is:

$H(s) = k \frac{ωₙ^2}{ s^2 + 2 \gamma ωₙ s + ωₙ^2}$

where ωₙ = 2π / tau, $\gamma$ is damping and k is gain.

SynthDiag.dribble — Function

dribble(
    data::Vector{Float64},
    decay_time_constant::Float64,
    fs::Float64,
)::Vector{Float64}

Function to model dribble effect when gas command falls too sharply due to remaining gas in the pipe. The function compares change in data(ΔD) at every time step with

$ΔD < -\frac{1}{τ f_s} e^{ - \left( \frac{1}{τ f_s} \right) }$

where τ(decay_time_constant) is the decay time constant and $f_s$(f_s) is the sampling frequency. If the above condition is found true, the function forces data to decay exponentially with the decay time constant untill the condition is false. The function returns the modified data.

SynthDiag.downsample_smooth — Function

downsample_smooth(
    tt::Vector{Float64},
    data::Vector{Float64},
    dt_res::Float64;
    default_start=0.0,
)::Vector{Float64}

Downsample and smooth the data to a new time resolution. The time axis does not need to be equally spaced in time. The function creates a new time axis with spacing given by dt_res and then averages the data points that fall within the new time bin. If no data points fall within the new time bin, the function uses the last data point to fill the new time bin. default_start is used to fill first time bin if no data is present in the first time bin.

downsample_smooth(
    tt::Vector{Float64},
    data::Vector{Float64},
    tt_res::Vector{Float64};
    default_start=0.0,
)::Vector{Float64}

Downsample and smooth the data to a new time resolution. The time axis does not need to be equally spaced in time. The function uses tt_res as the new resampled time axis and averages the data points that fall within the new time bin. If no data points fall within the new time bin, the function uses the last data point to fill the new time bin. default_start is used to fill first time bin if no data is present in the first time bin. Note that tt_res need not be equally spaced in time either.

SynthDiag.find_delay — Function

find_delay(
    data1::Vector{Float64},
    data2::Vector{Float64},
    dt::Float64,
)::Float64

Find the delay between two signals using cross-correlation. The function returns the delay in seconds. The function assumes that data1 is the reference signal and data2 is the signal that is delayed with respect to data1. The funciton assumes that data1 and data2 are sampled at the same rate and are of same length. dt is the time resolution of the data.

SynthDiag.get_gas_injection_response — Function

get_gas_injection_response(
    cmd::Vector{Float64},
    cmd_tt::Vector{Float64},
    P_ves::Vector{Float64},
    P_ves_tt::Vector{Float64},
    V_ves::Float64;
    resample_dt::Float64=0.02,
    gi_fit_guess::Vector{Float64}=[1.0, 1.0],
    response_curve_voltage::Vector{Float64}=collect(0:0.001:11),
)::Tuple{IMASDD.gas_injection__valve___response_curve, Dict{Symbol, Any}}

Function to fit a gas calibration shot data to a gas injection model. It requires 2 set of data:

cmd, and cmd_tt are the command sent in Volts and the corresponding time axis.

P_ves and P_ves_tt are the pressure in the vessel and the corresponding time axis.

V_ves is the volume of the vessel. The function returns a tuple of response_curve object and a valve model dictionary that can be used in compute_gas_injection! or compute_gas_injection functions.

Optional keyword arguments:

resample_dt: Time resolution to resample the data.

gi_fit_guess: Initial guess for the gas injection model parameters.

response_curve_voltage: Voltage axis for storing the response curve.

SynthDiag.gi_model — Function

gi_model(x::Vector{Float64}, p::Vector{Float64})::Vector{Float64}

Non linear gas injection model. The model is given by:

$\Gamma = p_1 (\sqrt{x^2 p_2^2 + 1} - 1)$

where x is the input voltage in Volts and $\Gamma$ is the flow rate in Pa m³/ s

SynthDiag.int_gi_model — Function

int_gi_model(x::Vector{Float64}, p::Vector{Float64})::Vector{Float64}

Cumulative sum of gas injection model. This function models the accumulated gas in the vessel volume. Note that this is a numerical cumulative sum and does not change the units of output. Returned output is in Pa m³/ s.

SynthDiag.get_required_gas_cmd — Function

get_required_gas_cmd(
        required_flow_rate::Float64,
        response_curve::IMASDD.gas_injection__valve___response_curve,
    )::Float64

Function to get the required gas command voltage to achieve the required flow rate.

get_required_gas_cmd(
    required_flow_rate::Vector{Float64},
    response_curve::IMASDD.gas_injection__valve___response_curve,
)::Vector{Float64}

Function to get the required gas command voltage to achieve the required flow rate.

Noise model

SynthDiag.Noise — Type

Noise

A structure to represent a noise model. It holds power spectral density of noise and/or the time series of noise. The power spectral density is stored as a DSP.Periodograms.Periodogram object.

Constructors:

Noise(time::Vector{Float64}, signal::Vector{Float64}, binwidth::Float64)

Create a noise model from a time series of noise. The power spectral density is calculated and stored for regenaeration of noise in future.

Noise(pgram::DSP.Periodograms.Periodogram)

Create a noise model from a power spectral density data given as a DSP.Periodograms.Periodogram object.

Noise(power_spectrum::Vector{Float64}, freq::AbstractRange)

Create a noise model from a power spectral density data given as a vector along with frequency values. The power spectral density is stored as a DSP.Periodograms.Periodogram calculated from the input data.

SynthDiag.generate_noise — Function

generate_noise(n::Noise, t::Vector{Float64})::Vector{Float64}

Generate a noise signal based on the power spectral density stored in the noise model that corresponds to the time axis t. The noise signal is generated by summing up the cosine waves with uniformly distributed random phase and amplitude based on the power spectral density.

SynthDiag.generate_noise! — Function

generate_noise!(n::Noise, t::Vector{Float64})::Noise

In-place version of generate_noise().